Mergesort ok

+++ include/line.hpp
@@ -1,7 +1,8 @@
+#include<cmath>
+#include"point.hpp"
+
 #ifndef LINE_H
 #define LINE_H
-
-#include"point.hpp"
 
 class Line 
 {
@@ -10,13 +11,13 @@ class Line
     Line(Point a, Point b);
     Point getB();
     Point getA();
-    int getPolarAngle();
+    double getPolarAngle();
     bool operator<(const Line& value);
     bool operator>(const Line& value);
   private:
     Point a;
     Point b;
-    int polarAngle;
+    double polarAngle;
     void calculePolarAngle();
 };
